examples/benchmark/requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
audioread==3.0.1
azure-core==1.31.0
azure-identity==1.19.0
babel==2.16.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bitsandbytes==0.43.3
bitstring==4.2.3
black==24.8.0
bleach==6.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==4.2.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-option-group==0.5.6
click-plugins==1.1.1
cmake==3.25.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
conformer==0.3.2
contourpy==1.3.0
contractions==0.1.73
cryptography==43.0.1
cycler==0.12.1
Cython==3.0.11
datasets==3.0.0
debugpy==1.8.7
decorator==5.1.1
deepspeed==0.14.2
defusedxml==0.7.1
diffusers==0.27.2
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
ecdsa==0.19.0
einops==0.8.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.16.1
fire==0.6.0
Flask==2.3.3
Flask-Cors==5.0.0
Flask-Executor==1.0.0
Flask-FontAwesome==0.1.5
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.65.0
gradio==4.32.2
gradio_client==0.17.0
grpcio==1.57.0
grpcio-tools==1.57.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
humanize==4.10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
immutabledict==4.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflate64==1.0.0
inflect==7.3.1
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==25.4.0
kiwisolver==1.4.7
langdetect==1.0.9
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.2.4
lightning-utilities==0.11.8
lit==15.0.7
llvmlite==0.43.0
loguru==0.7.2
loralib==0.1.2
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.15.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
modelscope==1.19.2
more-itertools==10.5.0
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.1.0
msrest==0.6.21
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
mypy-extensions==1.0.0
narwhals==1.12.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.16.0
openai==1.52.2
openai-whisper==20231117
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
optimum==1.22.0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passpy==1.0.2
pathspec==0.12.1
peft==0.13.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pooch==1.8.2
portalocker==2.10.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py7zr==0.22.0
pyahocorasick==2.1.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybcj==1.0.2
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pydash==8.0.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pynini==2.1.5
pynvml==11.5.3
pyparsing==3.2.0
pyperclip==1.9.0
pyppmd==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gnupg==0.5.2
python-json-logger==2.0.7
python-multipart==0.0.17
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
pyzstd==0.16.1
qa_metrics==0.2.17
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rich-click==1.8.3
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.7.1
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.14.1
SecretStorage==3.3.3
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
slam-llm==0.0.1
smmap==5.0.1
snac==1.2.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
sshpubkeys==3.3.1
stack-data==0.6.3
starlette==0.37.2
strictyaml==1.7.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
textsearch==0.0.24
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenize-rt==6.0.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.5.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.40.0
triton==3.0.0
typeguard==4.4.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.30.0
uvloop==0.21.0
wandb==0.18.1
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
WeTextProcessing==1.0.3
wget==3.2
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xdg==6.0.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.11.1
zipp==3.20.2
